refactor(app): split main.rs into focused modules (overlay, tray, detection_bridge)
main.rs grew to 375 lines doing 4 distinct things. Split by responsibility:

  main.rs (3 lines)         entrypoint
  lib.rs (~30 lines)        Tauri builder + setup wiring
  overlay.rs (~110 lines)   OverlayState + click-through cmds + cursor monitor + window positioning
  tray.rs (~40 lines)       tray icon + menu + handlers
  detection_bridge.rs       spawned task: GameEvent channel + enrichment + 2s refetch
    (~150 lines)

Pure refactor; no behavior change. The Legacy* IPC payload structs stay
in detection_bridge.rs for now and get removed in PR-B.
