+++ apps/desktop/app/src/tray.rs
@@ -1,0 +1,50 @@
+//! System tray icon + menu (Show/Hide, Quit) + click handlers.
+
+use std::sync::atomic::Ordering;
+use std::sync::Arc;
+use tauri::{
+    image::Image,
+    menu::{MenuBuilder, MenuItemBuilder},
+    tray::TrayIconBuilder,
+    AppHandle, Manager,
+};
+
+use crate::overlay::OverlayState;
+
+pub fn install(app: &AppHandle) -> tauri::Result<()> {
+    let toggle = MenuItemBuilder::with_id("toggle", "Show/Hide").build(app)?;
+    let quit = MenuItemBuilder::with_id("quit", "Quit").build(app)?;
+    let menu = MenuBuilder::new(app)
+        .item(&toggle)
+        .separator()
+        .item(&quit)
+        .build()?;
+
+    let _tray = TrayIconBuilder::new()
+        .icon(Image::from_bytes(include_bytes!("../icons/tray.png"))?)
+        .menu(&menu)
+        .tooltip("BrawlTome Overlay")
+        .on_menu_event(move |app: &AppHandle, event| {
+            match event.id().as_ref() {
+                "toggle" => {
+                    if let Some(w) = app.get_webview_window("overlay") {
+                        if w.is_visible().unwrap_or(false) {
+                            let _ = w.set_ignore_cursor_events(true);
+                            let st: tauri::State<Arc<OverlayState>> = app.state();
+                            st.clickthrough.store(true, Ordering::Relaxed);
+                            let _ = w.hide();
+                        } else {
+                            let _ = w.show();
+                        }
+                    }
+                }
+                "quit" => {
+                    app.exit(0);
+                }
+                _ => {}
+            }
+        })
+        .build(app)?;
+
+    Ok(())
+}
